+++ test/unit_tests/models/test_integration.py
@@ -15,9 +15,6 @@ from skorch.dataset import ValidSplit
 
 from braindecode.models.util import models_dict, models_mandatory_parameters
 from braindecode import EEGClassifier
-from braindecode.datasets import BaseDataset, BaseConcatDataset
-from braindecode.datasets.moabb import fetch_data_with_moabb
-from braindecode.preprocessing.windowers import create_windows_from_events
 
 # Generating the channel info
 chs_info = [dict(ch_name=f"C{i}", kind="eeg") for i in range(1, 4)]
@@ -161,45 +158,6 @@ def test_model_integration(model_name, required_params, signal_params):
         assert out.shape[:2] == (batch_size, sp["n_outputs"])
         # We add a "[:2]" because some models return a 3D tensor.
 
-
-bnci_kwargs = {
-    "n_sessions": 2,
-    "n_runs": 3,
-    "n_subjects": 9,
-    "paradigm": "imagery",
-    "duration": 3869,
-    "sfreq": 250,
-    "event_list": ("left", "right"),
-    "channels": ("C5", "C3", "C1"),
-}
-
-
-@pytest.fixture(scope="module")
-def concat_ds_targets():
-    raws, description = fetch_data_with_moabb(
-        dataset_name="FakeDataset", subject_ids=1, dataset_kwargs=bnci_kwargs
-    )
-
-    events, _ = mne.events_from_annotations(raws[0])
-    targets = events[:, -1] - 1
-    ds = [BaseDataset(raws[i], description.iloc[i]) for i in range(3)]
-    concat_ds = BaseConcatDataset(ds)
-    return concat_ds, targets
-
-
-@pytest.fixture(scope="module")
-def concat_windows_dataset(concat_ds_targets):
-    concat_ds, targets = concat_ds_targets
-    windows_ds = create_windows_from_events(
-        concat_ds=concat_ds,
-        trial_start_offset_samples=0,
-        trial_stop_offset_samples=0,
-        window_size_samples=750,
-        window_stride_samples=100,
-        drop_last_window=False,
-    )
-
-    return windows_ds
 
 
 @pytest.mark.parametrize(
